+++ NEWS.md
@@ -1,3 +1,5 @@
+# Rage (development version)
+
 # Rage 1.5.1
 
 * Patch to fix test errors on older systems (r-oldrel-macos-arm64). No changes to main package code.